apps/nagent/pyproject.toml
[project]
name = "nagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "nagent-core",
    "google-genai>=1.61.0",
    "html2text>=2025.4.15",
    "jsonref>=1.1.0",
    "jieba>=0.42.1",
    "jupytext>=1.19.1",
    "langchain-google-genai>=4.2.0",
    "matplotlib>=3.10.8",
    "paddlenlp>=2.5.2",
    "paddleocr>=3.3.3",
    "pillow>=12.1.0",
    "pymupdf>=1.26.7",
    "python-multipart>=0.0.22",
    "ragas>=0.4.3",
    "torch>=2.10.0",
    "unstructured[md,pdf]>=0.18.27",
]

[tool.uv.sources]
nagent-core = { workspace = true }

